requirements.txt
qdrant-client==1.9.1
psycopg2-binary==2.9.9
cohere==4.34
fastapi==0.104.1
python-dotenv==1.0.0
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pytest==7.4.3
pytest-cov==4.1.0
PyPDF2==3.0.1
beautifulsoup4
lxml
